Replace BitSlice<u64> with BitSlice<u32> for WASM compatibility
Changed all BitSlice<u64> to BitSlice<u32> throughout the codebase to fix
BitStore trait issues on 32-bit architectures. This involved:
- Updating all BitSlice type signatures
- Converting u64 data to u32 array views where needed
- Modifying DirectAtom and IndirectAtom bitslice methods
- Updating helper functions in zkvm-jetpack

All tests pass on 64-bit architecture.

+++ nockchain-core/crates/nockvm/rust/nockvm/src/interpreter.rs
@@ -1370,7 +1370,7 @@ fn mean_pop(stack: &mut NockStack) {
 
 fn edit(
     stack: &mut NockStack,
-    edit_axis: &BitSlice<u64, Lsb0>,
+    edit_axis: &BitSlice<u32, Lsb0>,
     patch: Noun,
     mut tree: Noun,
 ) -> Noun {
